harmonize async return class
