+++ src/api/index.ts
@@ -1,4 +1,4 @@
-export { Response, ResponseSig } from "./predictResponse";
+export { Response, ResponseSig } from "./documentResponse";
 export {
   Endpoint,
   CustomEndpoint,
@@ -7,4 +7,4 @@ export {
   STANDARD_API_OWNER,
   API_KEY_ENVVAR_NAME,
 } from "./endpoint";
-export { AsyncPredictResponse, ApiRequest, Job } from "./asyncPredictResponse";
+export { AsyncPredictResponse } from "./predictResponse";
